Fix note enums and constants to be across both initiatives
