+++ app/src/utils/enums/housing.ts
@@ -2,11 +2,6 @@
  * Housing initiative enums
  */
 
-export enum BringForwardType {
-  UNRESOLVED = 'Unresolved',
-  RESOLVED = 'Resolved'
-}
-
 export enum ContactPreference {
   PHONE_CALL = 'Phone call',
   EMAIL = 'Email',
@@ -20,13 +15,6 @@ export enum IntakeFormCategory {
   LOCATION = 'location',
   PERMITS = 'permits',
   APPLIED_PERMITS = 'appliedPermits'
-}
-
-export enum NoteType {
-  GENERAL = 'General',
-  BRING_FORWARD = 'Bring forward',
-  ENQUIRY = 'Enquiry',
-  ROADMAP = 'Roadmap'
 }
 
 export enum NumResidentialUnits {
